RECI-47 implement course page and create folder pages to hold all the pages
