+++ reciapp/lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:reciapp/today_special_page.dart';
-import 'home_page.dart';
+import 'package:reciapp/pages/today_special_page.dart';
+import 'pages/home_page.dart';
 
 void main() {
   runApp(const MyApp());
