+++ reciapp/lib/object/food.dart
@@ -1,10 +1,11 @@
-// ignore_for_file: prefer_const_literals_to_create_immutables
-
-import 'package:flutter/material.dart';
-
-class Food extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Container();
-  }
+class Food {
+  String image;
+  String name;
+  String author;
+  String content;
+  Food(
+      {required this.image,
+      required this.name,
+      required this.author,
+      required this.content});
 }
