+++ reciapp/lib/pages/course_page.dart
@@ -1,0 +1,174 @@
+// ignore_for_file: prefer_const_constructors
+
+import 'package:flutter/material.dart';
+import '../components/sidebar_menu.dart';
+import '../object/food_list.dart';
+import '../components/back_to_top_button.dart';
+import '../components/copyright.dart';
+import '../components/head_bar.dart';
+
+class CoursePage extends StatefulWidget {
+  const CoursePage({super.key});
+
+  @override
+  State<CoursePage> createState() => _CoursePageState();
+}
+
+class _CoursePageState extends State<CoursePage> {
+  ScrollController scrollController = ScrollController();
+  bool showbtn = false;
+
+  @override
+  void initState() {
+    scrollController.addListener(() {
+      //scroll listener
+      double showoffset =
+          10.0; //Back to top botton will show on scroll offset 10.0
+
+      if (scrollController.offset > showoffset) {
+        showbtn = true;
+        setState(() {
+          //update state
+        });
+      } else {
+        showbtn = false;
+        setState(() {
+          //update state
+        });
+      }
+    });
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      drawer: SideBarMenu(),
+      appBar: PreferredSize(
+        preferredSize: const Size.fromHeight(55),
+        child: HeadBar(),
+      ),
+      body: SingleChildScrollView(
+        controller: scrollController,
+        child: Column(
+          children: [
+            SizedBox(
+              height: MediaQuery.of(context).size.height * 0.09,
+              child: Stack(
+                children: [
+                  Positioned(
+                    child: Container(
+                      width: double.infinity,
+                      height: MediaQuery.of(context).size.height * 0.06,
+                      color: Color.fromARGB(255, 255, 231, 185),
+                    ),
+                  ),
+                  Positioned(
+                    top: 20,
+                    left: 20,
+                    child: Container(
+                        width: MediaQuery.of(context).size.width * 0.35,
+                        decoration: BoxDecoration(
+                            color: Colors.white,
+                            border:
+                                Border.all(color: Colors.black, width: 0.5)),
+                        child: Container(
+                          margin:
+                              EdgeInsets.symmetric(horizontal: 4, vertical: 2),
+                          child: Text(
+                            'Course',
+                            style: TextStyle(
+                              fontWeight: FontWeight.bold,
+                              fontSize: 18,
+                              fontFamily: 'Inter',
+                            ),
+                            textAlign: TextAlign.center,
+                          ),
+                        )),
+                  ),
+                ],
+              ),
+            ),
+            Row(
+              children: [
+                Container(
+                  decoration: BoxDecoration(
+                      border: Border(
+                          bottom: BorderSide(
+                              color: Theme.of(context).primaryColor,
+                              width: 0.8))),
+                  child: Text(
+                    'Breakfast',
+                    style: TextStyle(
+                      fontWeight: FontWeight.bold,
+                      fontSize: 18,
+                      fontFamily: 'Inter',
+                    ),
+                  ),
+                ),
+              ],
+            ),
+            SizedBox(height: MediaQuery.of(context).size.height * 0.01),
+            Container(
+              margin: EdgeInsets.symmetric(horizontal: 3),
+              height: MediaQuery.of(context).size.height * 0.35,
+              child: FoodList(),
+            ),
+            Row(
+              children: [
+                Container(
+                  decoration: BoxDecoration(
+                      border: Border(
+                          bottom: BorderSide(
+                              color: Theme.of(context).primaryColor,
+                              width: 0.8))),
+                  child: Text(
+                    'Lunch',
+                    style: TextStyle(
+                      fontWeight: FontWeight.bold,
+                      fontSize: 18,
+                      fontFamily: 'Inter',
+                    ),
+                  ),
+                ),
+              ],
+            ),
+            SizedBox(height: MediaQuery.of(context).size.height * 0.01),
+            Container(
+              margin: EdgeInsets.symmetric(horizontal: 3),
+              height: MediaQuery.of(context).size.height * 0.35,
+              child: FoodList(),
+            ),
+            Row(
+              children: [
+                Container(
+                  decoration: BoxDecoration(
+                      border: Border(
+                          bottom: BorderSide(
+                              color: Theme.of(context).primaryColor,
+                              width: 0.8))),
+                  child: Text(
+                    'Dinner',
+                    style: TextStyle(
+                      fontWeight: FontWeight.bold,
+                      fontSize: 18,
+                      fontFamily: 'Inter',
+                    ),
+                  ),
+                ),
+              ],
+            ),
+            SizedBox(height: MediaQuery.of(context).size.height * 0.01),
+            Container(
+              margin: EdgeInsets.symmetric(horizontal: 3),
+              height: MediaQuery.of(context).size.height * 0.35,
+              child: FoodList(),
+            ),
+          ],
+        ),
+      ),
+      floatingActionButton: BackToTopButton(scrollController, showbtn),
+      bottomNavigationBar: Copyright(),
+    );
+  }
+}
