+++ reciapp/lib/pages/login_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'components/copyright.dart';
+import '../components/copyright.dart';
 import 'home_page.dart';
 
 class LoginPage extends StatelessWidget {
